pyproject.toml
[tool.poetry]
name = "ravapp"
version = "0.1.0"
description = ""
authors = ["samolin <samolin.vlad@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.7"
djangorestframework = "^3.14.0"
python-dotenv = "^1.0.0"
psycopg2-binary = "^2.9.9"
httpx = "^0.25.2"
djangorestframework-simplejwt = "^5.3.0"
celery = "^5.3.6"
redis = "^5.0.1"
django-celery-beat = "^2.5.0"
fakeredis = "^2.20.0"


[tool.poetry.group.tests.dependencies]
pytest-django = "^4.7.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
